[Silver IV] Title: 짝수? 홀수?, Time: 0 ms, Memory: 2032 KB -BaekjoonHub

+++ 백준/Silver/11815. 짝수？ 홀수？/짝수？ 홀수？.cc
@@ -13,7 +13,7 @@ int main() {
         cin >> X;
 
 
-        long long r = (long long)floorl(sqrtl((long double)X));
+        long long r = (long long)floorl(sqrt((long double)X));
 
 
         if (r * r == X) {
